requirements.txt
dotenv 
ipykernel
openai 
requests
ipython
gradio
huggingface_hub
transformers
torch
bitsandbytes
accelerate
soundfile
google-genai
gradio